+++ modules/configuration/src/main/resources/db/migration/configuration/V15__allow_all_relevant_analysis_settings.sql
@@ -1,0 +1,6 @@
+ALTER TABLE configuration.monitoring_profiles
+    DROP CONSTRAINT IF EXISTS monitoring_profiles_analysis_minimum_matches_check;
+
+ALTER TABLE configuration.monitoring_profiles
+    ADD CONSTRAINT monitoring_profiles_analysis_minimum_matches_check
+        CHECK (analysis_minimum_matches IS NULL OR analysis_minimum_matches >= 0);